+++ src/resxEditor.ts
@@ -10,59 +10,65 @@ export class ResxEditor {
         this.context = context;
     }
 
-    public getHtmlForWebview(webview: vscode.Webview): string {
+    public getHtmlForWebview(webview: vscode.Webview, namespace: string, content: string): string {
 
         const scriptUri = webview.asWebviewUri(vscode.Uri.file(path.join(this.context.extensionPath, "out", "webpanelScript.js")));
         const styleUri = webview.asWebviewUri(vscode.Uri.file(path.join(this.context.extensionPath, "styles", "webpanel.css")));
-
+        const maPlusThick = webview.asWebviewUri(vscode.Uri.file(path.join(this.context.extensionPath, "styles", "ma-plus-thick.svg")));
+        const faPenToSquare = webview.asWebviewUri(vscode.Uri.file(path.join(this.context.extensionPath, "styles", "fa-pen-to-square.svg")));
+        const faRightLeft = webview.asWebviewUri(vscode.Uri.file(path.join(this.context.extensionPath, "styles", "fa-right-left.svg")));
         const nonce = getNonce();
 
         return `<!DOCTYPE html>
-        <html lang="en">
-        <head>
-            <meta charset="UTF-8">
-            <meta http-equiv="Content-Security-Policy"
-                content="default-src 'none'; img-src ${webview.cspSource}; style-src ${webview.cspSource}; script-src 'nonce-${nonce}';" />
-            <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-            <link href="${styleUri}" rel="stylesheet" />
-            <title>ResxFileName</title>    
-        </head>
-        <body>
-            <div id="container" class="topdiv">
-                <div id="leftThing">
-                    <button class="largeButtonStyle" id="addButton">Add New Resource</button>
-                </div>
-            
-                <div id="middleThing">
-                    <div id="diverr" class="error"></div>
-                </div>
-                
-                <div id="rightThing">
-                    <button class="smallButtonStyle" id="switchToEditor">Switch to Text Editor</button>
-                </div>
-            </div>
-            <table id="tbl">
-                <thead class="tableFixHead thead th">
-                    <th>Key</th>
-                    <th>Value</th>
-                    <th>Comment</th>
-                    <th> </th>
-                </thead>
-                <tbody>
-                </tbody>
-            </table>
-            <script nonce="${nonce}" src="${scriptUri}"></script>
-        </body>
-        </html>
-        `;
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="Content-Security-Policy"
+        content="default-src 'none'; img-src ${webview.cspSource}; style-src ${webview.cspSource}; script-src 'nonce-${nonce}';" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link href="${styleUri}" rel="stylesheet" />
+    <title>ResxFileName</title>
+</head>
+<body>
+    <div class="sticky-div">
+        <button id="addButton" class="btn primary">
+           <img src="${maPlusThick}" alt="Add Icon" class="icon filter-fefefe"> Add New Resource
+        </button>
+        
+        <button id="switchToTextEditorButton" class="btn secondary">
+            <img src="${faRightLeft}" alt="Switch Icon" class="icon filter-fefefe"> Switch to Text Editor
+        </button>
+        <div class="namespace-section">
+            <span>Namespace: <strong>${namespace}</strong></span>
+            <button id="changeNamespaceButton" class="btn secondary">
+                <img src="${faPenToSquare}" alt="Edit Icon" class="icon filter-fefefe"> Change Namespace
+            </button>
+        </div>
+        <p id="errorBlock" class="error-block"></p>
+    </div>
+
+    <table id="tbl">
+        <thead class="thead th">
+            <th>Key</th>
+            <th>Value</th>
+            <th>Comment</th>
+            <th> </th>
+        </thead>
+        <tbody>
+            ${content}
+        </tbody>
+    </table>
+    <script nonce="${nonce}" src="${scriptUri}"></script>
+</body>
+</html>`;
     }
 
     /**
      * Add a new key value back to text editor 
      */
     public addNewKeyValue(document: vscode.TextDocument, json: any) {
-        var newObj = JSON.parse(json);
-        var docDataList = ResxJsonHelper.getJsonData(document.getText());
+        const newObj = JSON.parse(json);
+        const docDataList = ResxJsonHelper.getJsonData(document.getText());
 
         var pos = docDataList.map((x) => { return x?._attributes?.name; }).indexOf(newObj._attributes.name);
 
@@ -96,7 +102,6 @@ export class ResxEditor {
         console.log("deleteKeyValue end");
         return this.updateTextDocument(document, JSON.stringify(currentData));
     }
-
 
     public updateTextDocument(document: vscode.TextDocument, dataListJson: any) {
         console.log("updateTextDocument start");