+++ src/webpanelMessageKind.ts
@@ -1,0 +1,8 @@
+export enum WebpanelPostMessageKind {
+    Switch = "switch",
+    Add = "add",
+    Delete = "delete",
+    Update = "update",
+    NamespaceUpdate = "namespace-update",
+    Alert = "alert"
+}